requirements.txt
asgiref==3.7.2
boto3==1.26.159
botocore==1.29.159
Django==4.2.2
django-environ==0.10.0
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
